% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IFA_Mode_Jumper}
\alias{IFA_Mode_Jumper}
\title{Exploratory Item Factor Analysis for Ordinal Response Data}
\usage{
IFA_Mode_Jumper(Y, M, gamma, Ms, sdMH, burnin, chain_length = 10000L)
}
\arguments{
\item{Y}{A N by J matrix of item responses.}

\item{M}{The number of factors.}

\item{gamma}{Mode-jumping tuning parameter. Man & Culpepper used 0.5.}

\item{Ms}{A vector of the number of score categories. Note 2 = two parameter normal ogive, >2 ordinal normal ogive.}

\item{sdMH}{A vector of tuning parameters of length J for implementing the Cowles (1996) Metropolis-Hastings threshold sampler.}

\item{burnin}{Number of burn-in iterations to discard.}

\item{chain_length}{The total number of iterations (burn-in + post-burn-in).}
}
\value{
A list that contains nsamples = chain_length - burnin array draws from the posterior distribution:
\itemize{
\item \code{LAMBDA}: A J by M by nsamples array of sampled loading matrices on the standardized metric.
\item \code{PSIs}: A J by nsamples matrix of vector of variable uniquenesses on the standardized metric.
\item \code{ROW_OUT}: A matrix of sampled row indices of founding variables for mode-jumping algorithm.
\item \code{THRESHOLDS}: An array of sampled thresholds.
\item \code{INTERCEPTS}: Sampled variable thresholds on the standardized metric.
\item \code{ACCEPTED}: Acceptance rates for mode-jumping Metropolis-Hastings (MH) steps.
\item \code{MHACCEPT}: A J vector of acceptance rates for item threshold parameters. Note that binary items have an acceptance rate of zero, because MH steps are never performed.
}
}
\description{
Implement the Man and Culpepper (2020) mode-jumping algorithm to factor analyze ordinal response data. Missing values should be specified as a non-numeric value such as NA.
}
\references{
Cowles, M. K. (1996), Accelerating Monte Carlo Markov chain convergence for cumulative link generalized linear models," Statistics and Computing, 6, 101-111.

Man, A. & Culpepper, S. A. (2020). A mode-jumping algorithm for Bayesian factor analysis. Journal of the American Statistical Association, doi:10.1080/01621459.2020.1773833.
}
\author{
Albert X Man, Steven Andrew Culpepper
}
