% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min2LL_ono}
\alias{min2LL_ono}
\title{Compute Deviance for Ordinal Probit Factor Model}
\usage{
min2LL_ono(N, J, Y, MISS, as, bs, theta, Ms, Kaps)
}
\arguments{
\item{N}{The number of observations.  (> 0)}

\item{J}{The number of items.  (> 0)}

\item{Y}{A N by J matrix of item responses.}

\item{MISS}{A N by J matrix of missing data indicators.}

\item{as}{A matrix of item loadings.}

\item{bs}{A vector of threshold parameters.}

\item{theta}{A matrix of factor scores.}

\item{Ms}{A vector of the number of score categories. Note 2 = two parameter normal ogive, >2 ordinal normal ogive.}

\item{Kaps}{A matrix of category thresholds.}
}
\value{
-2LL.
}
\description{
Internal function to compute -2LL using unstandardized parameters.
}
\author{
Steven Andrew Culpepper
}
