% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fitmeasures.R
\name{print.befa_fitmeasures}
\alias{print.befa_fitmeasures}
\title{Print Bayesian Fit Measures}
\usage{
\method{print}{befa_fitmeasures}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{befa_fitmeasures} object from \code{\link[=befa_fit_measures]{befa_fit_measures()}}.}

\item{digits}{Integer. Number of decimal places to display.}

\item{...}{Ignored.}
}
\description{
Displays a formatted table of Bayesian fit indices from a \code{befa_fitmeasures} object.
}
