% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proposal2}
\alias{proposal2}
\title{Rotate loadings and factor scores to a permuted positive lower triangle}
\usage{
proposal2(new_r_idx, lambda, factors)
}
\arguments{
\item{new_r_idx}{vector of row indices of length equal to the number of factors.}

\item{lambda}{Loading matrix.}

\item{factors}{A n x m matrix of factor scores that is rotated.}
}
\value{
A list of rotated loadings and factors.
}
\description{
Rotates loading matrix according to specified founding variable row indices.
}
\author{
Albert X Man
}
